Non-working version for demonstration of process.
At this point the code runs and outputs heatmap.dat to stdout.  Plotting this
heatmap with gnuplot (with the .gnu file provided) demonstrates that it has
converged to a steady-state where the upper and lower walls of the box have
T = 0 rather than dT/dy = 0.

We have discovered by a single run of our code that our code is not working
correctly and where it is.  Discovering this was easy because we set up
our code to output our data in a readable format /first/ and then wrote the
actual iteration code.  The last code I wrote was the boundary condition code
so it is fresh in my mind.